Fix return type
Reviewed-on: http://review.couchbase.org/91979
Tested-by: Build Bot <build@couchbase.com>

+++ spring/cbgen.py
@@ -81,7 +81,7 @@ def backoff(method: Callable, *args, **kwargs):
 
 
 @decorator
-def timeit(method: Callable, *args, **kwargs) -> int:
+def timeit(method: Callable, *args, **kwargs) -> float:
     t0 = time()
     method(*args, **kwargs)
     return time() - t0
